deploy swagger

+++ src/main/java/com/example/springquiz/repository/IAnswerRepository.java
@@ -4,6 +4,8 @@ import com.example.springquiz.model.domain.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface IAnswerRepository extends JpaRepository<Answer, Long> {
+public interface IAnswerRepository extends JpaRepository<Answer, Integer> {
 }
