+++ src/main/java/com/example/springquiz/repository/IQuestionRepository.java
@@ -4,6 +4,9 @@ import com.example.springquiz.model.domain.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface IQuestionRepository extends JpaRepository<Question, Long> {
+public interface IQuestionRepository extends JpaRepository<Question, Integer> {
+
 }
