+++ src/main/java/com/example/springquiz/repository/IQuizRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IQuizRepository extends JpaRepository<Quiz, Long> {
+public interface IQuizRepository extends JpaRepository<Quiz, Integer> {
 }
