+++ src/main/java/com/example/springquiz/service/IQuestionService.java
@@ -1,21 +1,20 @@
 package com.example.springquiz.service;
 
-import com.example.springquiz.model.dto.AnswerDTO;
 import com.example.springquiz.model.dto.QuestionDTO;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface IQuestionService {
-    Long createNewQuestion(QuestionDTO dto);
+    int createNewQuestion(QuestionDTO dto);
 
     List<Optional<QuestionDTO>> getAllQuestions();
 
-    Optional<QuestionDTO> getQuestionById(Long id);
+    Optional<QuestionDTO> getQuestionById(int id);
 
-    Optional<QuestionDTO> updateQuestion(Long id, QuestionDTO dto);
+    Optional<QuestionDTO> updateQuestion(int id, QuestionDTO dto);
 
-    void deleteQuestionById(Long id);
+    void deleteQuestionById(int id);
 
     void populate();
 }
