+++ src/main/java/com/example/springquiz/service/IQuizService.java
@@ -1,22 +1,21 @@
 package com.example.springquiz.service;
 
-import com.example.springquiz.model.dto.AnswerDTO;
-import com.example.springquiz.model.dto.QuestionDTO;
+
 import com.example.springquiz.model.dto.QuizDTO;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface IQuizService {
-    Long createNewQuiz(QuizDTO dto);
+    int createNewQuiz(QuizDTO dto);
 
     List<Optional<QuizDTO>> getAllQuizzes();
 
-    Optional<QuizDTO> getQuizById(Long id);
+    Optional<QuizDTO> getQuizById(int id);
 
-    Optional<QuizDTO> updateQuiz(Long id, QuizDTO dto);
+    Optional<QuizDTO> updateQuiz(int id, QuizDTO dto);
 
-    void deleteQuizById(Long id);
+    void deleteQuizById(int id);
 
     void populate();
 }
